backend/requirements.txt
Flask==3.1.1
flask-pymongo==2.3.0
python-dotenv==1.1.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.1.3
cryptography==42.0.8
dnspython==2.7.0